ISSUE-23: Fixes submit when for Mult. Associations
This Pull fixes #issue-23. Tags where not being saved, even when
correctly appended because the submit functions was iterating over all
associations for a DSID. Since the use case involved a very different
FORM, that had fields not present in all others, data was not being
passed to the XML document.
Also basic cleanup of iwi_set_form_element_byhash() was done to match
for `tag` the exact logic of `tab`and this way azure better
understanding.